Update CatHandler.js for NoSQL db

+++ handler/CatHandler.js
@@ -1,65 +1,80 @@
-import Cat from "../model/CatModel.js";
-import path from "path";
-import fs from "fs";
+import Cat from '../model/CatModel.js';
+import path from 'path';
+import fs from 'fs/promises';
 
 export const getCat = async (req, res) => {
     try {
-        const response = await Cat.findAll();
+        const response = await Cat.find();
         res.json(response);
     } catch (error) {
         console.log(error.message);
+        res.status(500).json({ error: 'Internal Server Error' });
     }
 }
 
 export const getCatById = async (req, res) => {
     try {
-        const response = await Cat.findOne({
-            where: {
-                id: req.params.id
-            }
-        });
+        const response = await Cat.findOne({ _id: req.params.id });
         res.json(response);
     } catch (error) {
         console.log(error.message);
+        res.status(500).json({ error: 'Internal Server Error' });
     }
 }
 
-export const saveCat = (req, res) => {
-    if (req.files === null) return res.status(400).json({ msg: "No File Uploaded" });
-    const name = req.body.name;
-    const age = req.body.age;
-    const gender = req.body.gender;
-    const size = req.body.size;
-    const coat = req.body.coat;
-    const breed = req.body.breed;
-    const description = req.body.description;
-    const file = req.files.image;
-    const fileSize = file.data.length;
-    const ext = path.extname(file.name);
-    const fileName = file.md5 + ext;
-    const url = `${req.protocol}://${req.get("host")}/images/${fileName}`;
-    const allowedType = ['.png', '.jpg', '.jpeg'];
-
-    if (!allowedType.includes(ext.toLowerCase())) return res.status(422).json({ msg: "Invalid Images" });
-    if (fileSize > 5000000) return res.status(422).json({ msg: "Image must be less than 5 MB" });
-
-    file.mv(`./public/images/${fileName}`, async (err) => {
-        if (err) return res.status(500).json({ msg: err.message });
-        try {
-            await Cat.create({ name: name, age: age, gender: gender, size: size, coat: coat, breed: breed, description: description, image: fileName, url: url });
-            res.status(201).json({ msg: "Cat Created Successfuly" });
-        } catch (error) {
-            console.log(error.message);
+export const saveCat = async (req, res) => {
+    try {
+        if (!req.files || Object.keys(req.files).length === 0) {
+            return res.status(400).json({ msg: 'No File Uploaded' });
         }
-    })
 
+        const { name, age, gender, size, coat, breed, description } = req.body;
+        const file = req.files.image;
+        const allowedTypes = ['.png', '.jpg', '.jpeg'];
+
+        const ext = path.extname(file.name);
+        if (!allowedTypes.includes(ext.toLowerCase())) {
+            return res.status(422).json({ msg: 'Invalid Image Type' });
+        }
+
+        const fileName = `${file.md5}${ext}`;
+        const url = `${req.protocol}://${req.get('host')}/images/${fileName}`;
+        const imagePath = `./public/images/${fileName}`;
+
+        if (file.size > 5000000) {
+            return res.status(422).json({ msg: 'Image must be less than 5 MB' });
+        }
+
+        file.mv(imagePath, async (err) => {
+            if (err) {
+                console.error(err);
+                return res.status(500).json({ msg: 'Error uploading file' });
+            }
+
+            await Cat.create({
+                name,
+                age,
+                gender,
+                size,
+                coat,
+                breed,
+                description,
+                image: fileName,
+                url,
+            });
+
+            res.status(201).json({ msg: 'Cat Created Successfully' });
+        });
+    } catch (error) {
+        console.log(error.message);
+        res.status(500).json({ error: 'Internal Server Error' });
+    }
 }
+
 
 export const updateCat = async (req, res) => {
     try {
-        const cat = await Cat.findOne({
-            where: { id: req.params.id },
-        });
+        const cat = await Cat.findById(req.params.id);
 
         if (!cat) {
             return res.status(404).json({ msg: 'No Data Found' });
@@ -86,10 +101,11 @@ export const updateCat = async (req, res) => {
             await fs.unlink(oldImagePath);
 
             const newImagePath = `./public/images/${fileName}`;
-            await file.mv(newImagePath);
+            await fs.writeFile(newImagePath, file.data);
 
             // Update the database with the new image information
-            await Cat.update(
+            await Cat.findByIdAndUpdate(
+                req.params.id,
                 {
                     name: req.body.name || cat.name,
                     age: req.body.age || cat.age,
@@ -100,16 +116,14 @@ export const updateCat = async (req, res) => {
                     image: fileName,
                     description: req.body.description || cat.description,
                     url: `${req.protocol}://${req.get('host')}/images/${fileName}`,
-                },
-                {
-                    where: { id: req.params.id },
                 }
             );
 
             res.status(200).json({ msg: 'Cat Updated Successfully' });
         } else {
             // Update other fields in the database even if no new image is provided
-            await Cat.update(
+            await Cat.findByIdAndUpdate(
+                req.params.id,
                 {
                     name: req.body.name || cat.name,
                     age: req.body.age || cat.age,
@@ -118,9 +132,6 @@ export const updateCat = async (req, res) => {
                     coat: req.body.coat || cat.coat,
                     breed: req.body.breed || cat.breed,
                     description: req.body.description || cat.description,
-                },
-                {
-                    where: { id: req.params.id },
                 }
             );
 
@@ -128,28 +139,30 @@ export const updateCat = async (req, res) => {
         }
     } catch (error) {
         console.error(error.message);
-        res.status(500).json({ msg: 'Internal Server Error' });
+        res.status(500).json({ error: 'Internal Server Error' });
     }
 };
 
-export const deleteCat = async(req, res)=>{
-    const cat = await Cat.findOne({
-        where:{
-            id : req.params.id
-        }
-    });
-    if(!cat) return res.status(404).json({msg: "No Data Found"});
+
+export const deleteCat = async (req, res) => {
+    const cat = await Cat.findById(req.params.id);
+
+    if (!cat) return res.status(404).json({ msg: "No Data Found" });
 
     try {
         const filepath = `./public/images/${cat.image}`;
-        fs.unlinkSync(filepath);
-        await Cat.destroy({
-            where:{
-                id : req.params.id
-            }
-        });
-        res.status(200).json({msg: "Cat Deleted Successfuly"});
+
+        // Check if the file exists before attempting to delete
+        if (fs.existsSync(filepath)) {
+            fs.unlinkSync(filepath);
+        } else {
+            console.log("File does not exist:", filepath);
+        }
+
+        await Cat.deleteOne({ _id: req.params.id });
+        res.status(200).json({ msg: "Cat Deleted Successfully" });
     } catch (error) {
         console.log(error.message);
+        res.status(500).json({ error: "Internal Server Error" });
     }
-}
+};
